requirements_test.txt
homeassistant==2023.7.3
pytest-homeassistant-custom-component==0.13.45
